Add high performance geolocation feature to haproxy
Those patches add high performance geolocation feature to HAProxy 1.4 and 1.5.
It provides an enhanced answer to that blog entry :
http://blog.exceliance.fr/2012/07/02/use-geoip-database-within-haproxy/

A new keyword is available :
  geolocate <src|hdr_ip(<header>,<occ>) <filename>
where <filename> points to a file containing lines described as :
  <ip or CIDR network> <value>

A new ACL is also available :
  geoloc <values>
where <values> is one of the value found in the file, or "*" to match any of
them.

A token is introduced for redirection : ${geoloc}. This token is substituted by
the value that matched the geolocation.

Usage example :
  geolocate hdr_ip(X-Forwarded-For,-1) geolocation.txt
  redirect location http://{$geoloc}.website.com/ if { geoloc fr us }
  redirect prefix /${geoloc} if { geoloc * }