FIX: Use haproxy sample instead of acl
haproxy 1.5-dev18 was modified to allow the direct use of sample fetches
instead of acls [1].
Between 1.5-dev18 and 1.5-dev19, acls were cleaned to remove useless ones by
replacing them with fetches [2].

The patch needs to be updated to follow thoses modifications in order to
initialize the geolocation feature, otherwise a fatal error will be displayed
(the code tried to find the "src" acl when initializing). Now, we're using the
fetch equivalent.

[1] http://haproxy.1wt.eu/git?p=haproxy.git;a=commit;h=bef91e7144efe6ce65f627c0ae94af2e7593f7f5
[2] http://haproxy.1wt.eu/git?p=haproxy.git;a=commit;h=6d4e4e8dd2c6cb3af4ffffe552821d131fb7290d